deepseek/rest/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
structlog==23.2.0
orjson==3.9.10
PyJWT==2.8.0
